API

TuePlots.SETTINGS — Constant

Dict of available TuePlots settings.

Use with MakieCore.Theme or get_plotsjl_theme_kwargs.

TuePlots.get_available_settings — Function

get_available_settings()

Return a list of settings available in TuePlots.jl.

Effectively this function just returns the keys of SETTINGS.

MakieCore.Theme — Type

MakieCore.Theme(setting::TuePlotsSetting;
    font = true, fontsize = true, figsize = :full, thinned = true

Make a Makie Theme out of the TuePlotsSetting.

TuePlots.get_plotsjl_theme_kwargs — Function

get_plotsjl_theme_kwargs(setting::TuePlotsSetting;
    font = true, fontsize = true, figsize = :full, thinned = trues+

Build a named tuple to be used with Plots.theme out of the TuePlotsSetting.

Misc

TuePlots.MakieThinTheme — Constant

Makie theme with thinner lines everywhere.

Helpful to improve most Makie plots, since the default settings are often chosen for much larger plots than the small figure sizes used in papers.